+++ Agora-Unreal-SDK-CPP-Example/Source/AgoraExample/CustomCaptureAudioWidget.cpp
@@ -80,7 +80,7 @@ void UCustomCaptureAudioWidget::onJoinChannelSuccess(const char* channel, agora:
 {
 	AsyncTask(ENamedThreads::GameThread, [=]()
 	{
-		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UCustomCaptureAudioWidget::JoinChannelSuccess uid: %d"), (int64)uid));
+		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UCustomCaptureAudioWidget::JoinChannelSuccess uid: %u"), uid));
 	});
 }
 
@@ -88,7 +88,7 @@ void UCustomCaptureAudioWidget::onUserJoined(agora::rtc::uid_t uid, int elapsed)
 {
 	AsyncTask(ENamedThreads::GameThread, [=]()
 	{
-		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UCustomCaptureAudioWidget::onUserJoined uid: %d"), (int64)uid));
+		GEngine->AddOnScreenDebugMessage(-1, 30.f, FColor::Blue, FString::Printf(TEXT("UCustomCaptureAudioWidget::onUserJoined uid: %u"), uid));
 	});
 }
 #pragma endregion RtcEngineCallBack
